requirements.txt
streamlit
google-generativeai
chromadb
python-dotenv
pypdf
tenacity
